Creates a separate module for the hibernate backend.
