+++ client-misk-hibernate/README.md
@@ -1,0 +1,3 @@
+# client-misk-hibernate
+
+Hibernate backfila client backend implementation for misk apps. This uses the misk-hibernate Query to define the entries to backfill over.
