+++ client-misk-hibernate/build.gradle
@@ -6,27 +6,24 @@ dependencies {
   implementation dep.moshiKotlin
   implementation dep.wireRuntime
   implementation dep.guice
-  implementation dep.okHttp
-  implementation dep.retrofit
-  implementation dep.retrofitMock
-  implementation dep.retrofitMoshi
-  implementation dep.retrofitWire
   implementation dep.okio
   implementation dep.kotlinStdLib
   implementation dep.wireMoshiAdapter
 
-  api project(':client')
+  api project(':client-misk')
   testApi project(':client-misk-testing')
 
   implementation dep.misk
+  api dep.miskHibernate
 
   testImplementation dep.assertj
   testImplementation(dep.miskTesting)
+  testImplementation(dep.miskHibernateTesting)
   testImplementation dep.junitEngine
 }
 
 jar {
-  baseName = "backfila-client-misk"
+  baseName = "backfila-client-misk-hibernate"
 }
 
 afterEvaluate { project ->
