+++ client-misk-hibernate/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=client-misk-hibernate
+POM_NAME=client-misk-hibernate
+POM_DESCRIPTION=Misk hibernate backfila client backend implementation library
+POM_PACKAGING=jar
