+++ client-misk-hibernate/src/main/kotlin/app/cash/backfila/client/misk/hibernate/internal/HibernateBackfillOperator.kt
@@ -3,9 +3,9 @@ package app.cash.backfila.client.misk.hibernate.internal
 import app.cash.backfila.client.misk.NoParameters
 import app.cash.backfila.client.misk.hibernate.HibernateBackfill
 import app.cash.backfila.client.misk.hibernate.PkeySqlAdapter
-import app.cash.backfila.client.misk.internal.BackfilaParametersOperator
-import app.cash.backfila.client.misk.spi.BackfillOperator
 import app.cash.backfila.client.misk.internal.BackfillOperatorFactory
+import app.cash.backfila.client.misk.spi.BackfilaParametersOperator
+import app.cash.backfila.client.misk.spi.BackfillOperator
 import app.cash.backfila.protos.clientservice.GetNextBatchRangeRequest
 import app.cash.backfila.protos.clientservice.GetNextBatchRangeResponse
 import app.cash.backfila.protos.clientservice.GetNextBatchRangeResponse.Batch
