+++ client-misk-hibernate/src/test/kotlin/app/cash/backfila/client/misk/hibernate/NoParametersTest.kt
