+++ client-misk-hibernate/src/test/resources/extra_my.cnf
