+++ settings.gradle
@@ -1,4 +1,5 @@
 include 'client'
 include 'client-misk'
+include 'client-misk-hibernate'
 include 'client-misk-testing'
 include 'service'
